refactor: Move device creation into helper function
Move children device creation out from `Nexus::new_child` into a helper
function. Create the children devices before the Nexus is created and
add them after the Nexus's initialization.

This is in preperation for zoned storge support, as we need to inspect
the children device features before creating a Nexus such that we can
decide whether the Nexus must be zoned or not.

Signed-off-by: Dennis Maisenbacher <dennis.maisenbacher@wdc.com>