feat(nexus): Add features for Zoned Storage Support
Adding setter and getter for zone storage related fields to the Nexus and
BlockDevice.

Inspect the children before creating the Nexus to decide whether the Nexus
needs to be zoned or not. The Nexus inherrits the zoned related fields
form its child.

For now a zoned Nexus does not allow replication.

Partitioning of a zoned Nexus is also not supported for now.
The user will optain the whole zoned Nexus if requested.

Signed-off-by: Dennis Maisenbacher <dennis.maisenbacher@wdc.com>